Updated vault id in auction section
